tests/submission/W3C/domhtml/HTMLDocument11.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - DOCUMENT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-document-anchors" rel="help">
<link rel="author" title="NIST">
</head>
<body id="TEST-BODY"> 
<h1>This is a Document</h1>
<p>This is a paragraph.</p>
<img alt="Yellow image" height="50" src="/support/images/yellow-rectangle.png" width="100">
<p class="copyright"><a href="http://www.csail.mit.edu/"><acronym title="Massachusetts Institute of Technology">MIT</acronym></a>, <a href="http://www.ercim.eu/"><acronym title="European Research Consortium for Informatics and Mathematics">ERCIM</acronym></a>,
<a href="http://www.keio.ac.jp/" name='foo'>Keio</a></p>


<script>
var domtest = async_test("The anchors attribute must return an HTMLCollection rooted at the Document node, whose filter matches only a elements with name attributes.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vanchors;
      var vlength;
      var doc;
      doc = window.document;
      vanchors = doc.anchors;

      vlength = vanchors.length;

      assert_equals(vlength,1);
       
    domtest.done();
}));

</script>
<p>The anchors attribute must return an HTMLCollection rooted at the Document node, whose filter matches only a elements with name attributes.</p>
<div id="log"></div>
</body></html>
